Junit Writing

+++ ExcelToDatabaseTask/src/main/java/com/example/demo/entity/EmployeeRepository.java
@@ -2,6 +2,6 @@ package com.example.demo.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<EntityClass, Integer> {
+public interface EmployeeRepository extends JpaRepository<EntityClass, Long> {
 
 }
